backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-dotenv
